use size_t for fk_skiplist_t.len again

+++ fk_skiplist.h
@@ -21,7 +21,7 @@ typedef struct {
 typedef struct {
     fk_skipnode_t *head;
     int level;/* the max level of the nodes */
-    uint64_t len;
+    size_t len;
     fk_skipnode_op *skop;
 } fk_skiplist;
